requirements_client.txt
anyio==4.2.0
attrs==23.2.0
certifi==2023.11.17
charset-normalizer==3.3.2
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
environs==11.0.0
fonttools==4.47.0
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
kiwisolver==1.4.5
Logbook==1.7.0.post0
marshmallow==3.21.2
matplotlib==3.8.2
ntplib==0.4.0
numpy==1.26.3
packaging==23.2
pillow==10.2.0
pyorbital==1.8.1
pyparsing==3.1.1
PyQt5==5.15.10
PyQt5-Qt5==5.15.13
PyQt5-sip==12.13.0
python-dateutil==2.8.2
python-dotenv==1.0.1
PyYAML==6.0.1
Represent==2.0
requests==2.31.0
rush==2021.4.0
scipy==1.11.4
six==1.16.0
sniffio==1.3.0
spacetrack==1.2.0
tqdm==4.66.1
urllib3==2.1.0
